+++ softgroup/basis/common-router/src/main/java/com/softgroup/common/router/api/RouterHandler.java
@@ -5,4 +5,5 @@ import com.softgroup.common.protocol.Request;
 public interface RouterHandler extends Handler {
 
 	String getRouteKey(final Request<?> msg);
+
 }
