+++ softgroup/services/authorization/authorization-impl/main/java/com/softgroup/authorization/impl/handler/AuthorizationRouterHandlerLogin.java
@@ -1,18 +1,22 @@
-package com.softgroup.authorization.api.router;
+package com.softgroup.authorization.impl.handler;
 
+import com.softgroup.authorization.api.router.AuthorizationRouterHandler;
 import com.softgroup.common.protocol.Request;
 import com.softgroup.common.protocol.Response;
+import org.springframework.stereotype.Component;
 
 /**
  * Created by polev on 25.02.2017.
  */
-public class AuthorizationRouterHandlerRegister implements AuthorizationRouterHandler {
+
+@Component
+public class AuthorizationRouterHandlerLogin implements AuthorizationRouterHandler {
     public String getRouteKey(Request<?> msg) {
         return null;
     }
 
     public String getName() {
-        return "register";
+        return "login";
     }
 
     public Response<?> handle(Request<?> msg) {
