+++ softgroup/services/pom.xml
@@ -14,6 +14,7 @@
 	<modules>
 
 		<module>authorization</module>
+		<module>profile</module>
 	</modules>
 
 	<properties>
